+++ src/main/java/me/drex/vanillapermissions/mixin/selector/EntitySelectorMixin.java
@@ -18,8 +18,8 @@ public abstract class EntitySelectorMixin {
             //? if > 1.21.10 {
             target = "Lnet/minecraft/server/permissions/PermissionSet;hasPermission(Lnet/minecraft/server/permissions/Permission;)Z"
             //? } else if >= 1.21.6 {
-//            target = "Lnet/minecraft/commands/CommandSourceStack;allowsSelectors()Z"
-            //?} else {
+            /*target = "Lnet/minecraft/commands/CommandSourceStack;allowsSelectors()Z"
+            *///?} else {
             /*target = "Lnet/minecraft/commands/CommandSourceStack;hasPermission(I)Z"
             *///?}
         )